uir rounding change

+++ crate/src/exec/basic/statistics.rs
@@ -1,21 +1,21 @@
-pub fn mean(data: &[i32]) -> Option<f32> {
-    let sum = data.iter().sum::<i32>() as f32;
+pub fn mean(data: &[i64]) -> Option<f64> {
+    let sum = data.iter().sum::<i64>() as f64;
     let count = data.len();
 
     match count {
-        positive if positive > 0 => Some(sum / count as f32),
+        positive if positive > 0 => Some(sum / count as f64),
         _ => None,
     }
 }
 
-pub fn std_deviation(data: &[i32]) -> Option<f32> {
+pub fn std_deviation(data: &[i64]) -> Option<f64> {
     match (mean(data), data.len()) {
         (Some(data_mean), count) if count > 0 => {
             let variance = data.iter().map(|value| {
-                let diff = data_mean - (*value as f32);
+                let diff = data_mean - (*value as f64);
 
                 diff * diff
-            }).sum::<f32>() / count as f32;
+            }).sum::<f64>() / count as f64;
 
             Some(variance.sqrt())
         },
